move debug reporter to its own file
